Extract ConfigurationCacheDirIntegrationTest
from ConfigurationCacheIntegrationTest
